update to moments

+++ scripts/copy_media.py
@@ -21,16 +21,15 @@ the last parameter is a translate, filter option
 the prefix in place in all source logs is the first option
 and the local destination is the second option
 
-TODO:
-convert to using medialist.sources
 """
 
 import sys, os, re
 import subprocess
-from moments.journal import Journal, load_journal
+from moments.journal import Journal
+from moments.path import load_journal
 from moments.association import check_ignore
 #from medialist.medialist import MediaList
-from medialist.sources import Converter, Sources
+from moments.sources import Converter, Sources
 
 def flatten_structure(source):
     destination = '/binaries/music/vinyl/flat/'
@@ -62,8 +61,9 @@ def make_destination(source, translate):
     else:
         (pre, post) = ('/c/media', '')
 
+    #print "source: %s" % source
     destination = source.replace(pre, post)
-    print "destination: %s" % destination
+    #print "destination: %s" % destination
     
     #change the path used in logs to your local path
     #in this case, just remove the prefix /c/media
@@ -76,16 +76,19 @@ def make_destination(source, translate):
     
     return destination
 
-def copy_file(source, destination):
+def copy_file(source, destination, verbose=False):
     """
     copy a single file 
     """
     #cp = subprocess.Popen("cp %s %s" % (source, destination), shell=True, stdout=subprocess.PIPE)
     #cp.wait()
     command = 'rsync -av "%s" "%s"' % (source, destination)
-    print command
     rsync = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    print rsync.communicate()[0]
+    if verbose:
+        print command
+        print rsync.communicate()[0]
+    else:
+        rsync.communicate()[0]
 
 
 def copy_files(journal, translate=None):
@@ -96,27 +99,31 @@ def copy_files(journal, translate=None):
     
     #j = Journal()
     #j.from_file(journal)
-    j = load_journal(journal)
+    #j = load_journal(journal)
     #m = MediaList()
     #m.from_journal(j, local_path='/c')
-    sources = Sources()
-    converter = Converter(sources)
-    converter.from_entries(j)
+    #sources = Sources()
+    converter = Converter()
+    sources = converter.from_journal(journal)
     for i in sources:
-        print i
+        #print i
         if re.search('\.mp3', i.path):
-            destination = make_destination(i.path, translate)
-            #print "SOURCE: %s" % i
-            #print "DEST: %s" % destination
+            if os.path.exists(i.path):
+                destination = make_destination(i.path, translate)
+                #print "SOURCE: %s" % i
+                #print "DEST: %s" % destination
 
-            if os.path.exists(destination):
-                print "skipping: %s" % destination
+                if os.path.exists(destination):
+                    print "skipping: %s" % destination
+                    pass
+                else:
+                    dest_path = os.path.dirname(destination)
+                    if not os.path.exists(dest_path):
+                        os.makedirs(dest_path)
+                        
+                    copy_file(i.path, destination)
             else:
-                dest_path = os.path.dirname(destination)
-                if not os.path.exists(dest_path):
-                    os.makedirs(dest_path)
-
-                copy_file(i.path, destination)
+                print "COULD NOT FIND FILE: %s" % i.path
     
 def main():
     if len (sys.argv) > 1:
